Not all the GetTypeName calls had the option to DoNotThrow specified.
If we don't check that the Type names can resolve in the Match method then it will be too late when the Resolution parts execute and they return with null.

So I added checks in the Match method to ensure it can resolve the Repository, Domain entity and DTO instances.
AB#989