Fixed: Creating a new Clean Architecture application would not compile off the bat.
This is because the old interop used to add domain events to entities, but that logic has now moved to the domain events module itself. Because the template wouldn't refer directly to the interop module, it was installing the old version via the dependency here, meaning we would get double generation of the relevant logic on entities.