CRUD Module:
When a Query By Id is set to return a nullable object and has a repository that can't find the entity, it will return `default` as opposed to throw a `NotFoundException`.

Unit Test Module:
Updated test for Query By Id cases where a nullable object is returned, it shouldn't check for a `NotFoundException` to be thrown but rather to assert that the result is `default`.