Optimized Etag retrieval method
The 'GetEtag' method has been refactored to only be added when optimistic concurrency is in use. This change reduces unnecessary code and improves performance by avoiding the execution of this method when it's not needed. The 'GetEtag' method was also removed from several test repositories as part of this optimization.
"Refactored Etag retrieval for optimistic concurrency

The 'GetEtag' method has been conditionally added only when optimistic concurrency is in use. This change reduces unnecessary code execution and improves performance. The 'GetEtag' method was also removed from several test repositories as part of this optimization."