Updated module versions and fixed package references
The commit includes updates to the version numbers of several modules. The changes are mainly in the .imodspec files where the version number has been incremented. Additionally, there's a change in NuGetPackages.cs file where the MicrosoftExtensionsHttp method now returns an instance of NugetPackageInfo instead of NugetPackageInfo directly. Also, it corrects the version number for when GetMaxNetAppVersion() returns (5, 0). Lastly, a reference to 'Microsoft.Extensions.Http' in a test project file was downgraded from version "8.0.0" to "7.0.0".