Updated module version and fixed base type application
The module version has been updated to 5.0.3-pre.0 in both the imodspec file and csproj file. A bug fix was also implemented where the Data Contract now correctly applies the Base Type when it is set in the designer. This change is reflected in the release notes as well as in the DataContractTemplatePartial.cs file, where a condition was added to check if Model.BaseType is not null before applying it.