Updated module version and fixed cloud events generation
The module version has been updated to 2.0.5-pre.1 from 2.0.4 in the imodspec file. A bug fix was implemented where `UseCloudEvents` will no longer be generated in the wrong location when using Minimal Hosting Model, by publishing an ApplicationBuilderRegistrationRequest with a priority of -200 instead of directly inserting it into the statements list.