Refactored DTOValidator and ValidationRulesExtensions
The code has been refactored to improve the handling of advanced mappings in DTO validation. A new method, GetAdvancedMappings, was introduced to encapsulate the logic for retrieving advanced mappings. This method is now used in multiple places, reducing code duplication. The ValidationRulesExtensions class was also updated to accept an additional parameter for associated elements, which allows it to handle advanced mappings more accurately.